LayoutTests/http/tests/contentextensions/service-worker-block-everything-if-localhost.html

<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
function waitForState(worker, state)
{
    if (worker.state === state)
        return Promise.resolve(state);

    return new Promise(function(resolve) {
      worker.addEventListener('statechange', function() {
          if (worker.state === state)
            resolve(state);
        });
    });
}

promise_test(async t => {
    const registration = await navigator.serviceWorker.register("service-worker-block-everything-if-localhost-serviceworker.js", {});
   await waitForState(registration.installing, "activated");

   registration.active.postMessage("claim");
   await new Promise(resolve => navigator.serviceWorker.onmessage = resolve);
}, "Setup service worker");

promise_test(async t => {
    assert_true(!!navigator.serviceWorker.controller);
    return fetch("/security/resources/abe-allow-star.py").then(assert_unreached, () => {});
}, "Validate that localhost load is blocked");

promise_test(async t => {
    if (window.internals)
        internals.disableContentExtensionsChecks();

    return fetch("/security/resources/abe-allow-star.py").then(() => {}, assert_unreached);
}, "Validate that localhost load is not blocked if content extension is off");
</script>
</body>
</html>
